[PROD-8071] Use the customer tenant ID in the engine containers
In the context of a managed app deployed in the customer tenant,
the AMQP Consumer gets rejected because the JWT Token issuer is
the core Platform Tenant.
Fixing this requires passing the customer tenant ID instead, which requires
the customer to consent to the core App Registration (magic Enterprise Application link).
And this might be incompatible with the 'stack steps' feature,
because the 'engine' step needs a dedicated tenant ID environment variable.